+++ Packages/com.pinchasvaknin.arenaforge/Tests/EditMode/RoadKerbTests.cs
@@ -118,7 +118,7 @@ namespace ArenaForge.Tests
             "ca395703f432dc16", "19a237429cd0b3e9", "c2b78337838ff6cf", "9090dd353132e4e5",
             "493ae577cfa0bc54", "4ba0f36a7409309f", "0fb691008c7b9418", "39285c02466a4d96",
             "3cb92e865d5fe6b7", "e6fbfd9cd17acc78", "fb7101541e806682", "690fa65410a9853b",
-            "a13ae06087fa8f4b", "cbadfbb7c93313a1", "0c2da338ca098560", "70cfe64e02686c2a",
+            "afe6fb981cc751c6", "cbadfbb7c93313a1", "0c2da338ca098560", "70cfe64e02686c2a",
             "a41d50dfb87ff4fd", "d4f752cad77d1601", "9661f6e5d33774e2", "df478c929c473a15",
             "d9c611c66381b705", "cb40d9acfc2e2dbc", "eafc3dfef80829cf", "e28a4c12cbea8aed",
             "d07d4fd630d76bd3", "03c2a529402f6141", "6b236166b9ee3fc7", "dd732bf654029392",